+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Release Notes for Craft Commerce
 
+
+### Fixed
+- Fixed a bug where disabling all variants throws an error on Edit Product Page.
+
 ## 4.0.4 - 2022-06-22
 
 > {note} If you’ve already upgraded a site to Commerce 4, please go to **Commerce** → **Promotions** → **Discounts** and review your discounts’ coupons’ Max Uses values, as the `commerce/upgrade` command wasn’t migrating those values properly before this release.